Fix error

+++ src/Weather_site.Core/Entities/Main.cs
@@ -14,7 +14,7 @@ namespace Weather_site.Core.Entities
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public double temp { get; set; }
         public double feels_like { get; set; }
-        public int temp_min { get; set; }
+        public double temp_min { get; set; }
         public double temp_max { get; set; }
         public int pressure { get; set; }
         public int humidity { get; set; }
